Added additional documentation to remaining files and organized code
