+++ lib/arcane_framework.dart
@@ -1,14 +1,14 @@
 library arcane_framework;
 
 export "package:app_tracking_transparency/app_tracking_transparency.dart";
-export "package:arcane_framework/arcane.dart";
+export "package:arcane_framework/src/arcane.dart";
 export "package:arcane_framework/src/arcane_app.dart";
-export "package:arcane_framework/src/authentication/authentication_service.dart";
-export "package:arcane_framework/src/feature_flags/feature_flags_service.dart";
-export "package:arcane_framework/src/id/id_service.dart";
-export "package:arcane_framework/src/logging/logging_service.dart";
 export "package:arcane_framework/src/providers/environment_provider.dart";
 export "package:arcane_framework/src/providers/service_provider.dart";
-export "package:arcane_framework/src/reactive_theme/reactive_theme_service.dart";
-export "package:arcane_framework/src/secure_storage/secure_storage_service.dart";
+export "package:arcane_framework/src/services/authentication/authentication_service.dart";
+export "package:arcane_framework/src/services/feature_flags/feature_flags_service.dart";
+export "package:arcane_framework/src/services/id/id_service.dart";
+export "package:arcane_framework/src/services/logging/logging_service.dart";
+export "package:arcane_framework/src/services/reactive_theme/reactive_theme_service.dart";
+export "package:arcane_framework/src/services/secure_storage/secure_storage_service.dart";
 export "package:result_monad/result_monad.dart";
